Remove conversions to data frames
For the input data `x`, we're not trying to respect the class of the
input value to accommodate other data types such as sparse matrices.

There are some complications though; for pre-processing we throw a
warning if the data are not simple matrices or data frames since there
is some infrastructure that does not exist for other classes( e.g.
`complete.cases`). We also throw a warning if `returnData <- TRUE` and
it cannot be converted to a data frame.