fixed a bug in finding the reference function
I pass the points of the curve to the `approx` function to see where
the x-axis point is since the reference line probably falls in-between
points on the lift curve. For whatever reason, for these data,`approx`
produces a wildly inaccurate number. However, if I just give that
function the two points above and below the reference value, it works.