Make mnLogLoss() more efficient
Previously, the function constructed the predicted probability of the
observed outcome by building up a matrix of dummy variables and
multiplying with the matrix of predicted probabilities.  Now it uses
match() to identify the column index corresponding to the observed
class for each observation, then uses the matrix form of `[` to extract
the relevant probabilities.

Results in a speedup of roughly 10x, per back-of-the-envelope tests.