call make.names on obsLevels to properly handle spaces in factors
If the response is a factor which has levels which contain spaces, the
out object from extractProb() will have called make.names() to change
the column names to have periods instead of spaces.

This results in

Error in `[.data.frame`(out, , obsLevels, drop = FALSE) :
  undefined columns selected

when predict is called on an object.

An eventual real fix is to set check.names=FALSE in the right bit of
extractProb() which is having this effect.