Tuning grid fix
eta parameter is primarily dependent on nrounds in the grid. It prevents "overfitting" but you require more rounds if you intend to fiddle with eta (as the shrinkage is huge) - typical eta parameters for maximum optimization using xgbTree is below 0.001. eta should be put as close to 0 as possible in order to get the maximum from an optimized parameter set. It is invariant to any other parameters than nrounds.
Expanding the grid using nrounds + eta parameter is theoretically inconsistent as eta is the last parameter to look at, not one of the first parameters to look at. Good values for trying to find the optimal settings is usually 0.1 if you got a large amount of time, or 0.3 if you want to make it fast.

Colsample_bytree also affects nrounds. The higher Colsample_bytree is, the higher nrounds you need. However, if you reach local CV optimum quickly, it should stop immediately at a specific round (using early.stop.round preferably, but it could take very long depending on the dataset size - here there is no early.stop.round). Here, it is not a good idea. Therefore, it is useful to remove eta parameter optimization and use colsample_bytree parameter optimization, as it can yield a higher optimization for performance metrics. Some data sets are converging with a very good metric when using 1.0, while a lot of some data sets are converging better with smaller values.

max_depth -> higher number leads to a lower amount of nrounds as it changes the tree structure (convergence is slower - there are very rare cases where it's slower to converge using smaller depth)
gamma -> higher number leads to a higher amount of nrounds (convergence is slower due to train performance metric = test performance metric when gamma gets higher)
colsample_bytree -> higher number leads to a higher amount of nrounds (due to a potentially lower access to useful features)
eta -> variant on nrounds, if eta is lower, then nrounds must be forcibly higher due to higher weight shrinkage (slower convergence when eta is low)
min.child.weight -> overfitting control parameter, same as gamma but it is mainly linked with tree depth (pruning trees whose terminal nodes have a specific lower hessian weight)

eta = 0.3 -> default value of xgbTree, dependent on nrounds - fiddling eta with a so small amount of nrounds will not yield any interesting output as higher eta will pick up convergence faster. It should be left to the user to optimize eta/nrounds, as messing the pair together has a very large effect on any other parameters.

It is why it does not make sense to optimize eta along other variables, as it is the last parameter to look at. Typical xgb.train users are using xgb.cv with an unrealistic amount of nrounds and early.stop.round to find the best nrounds count. They optimize the parameters (except eta), then when they want to get the highest performance they drop eta to an abysmal value (like 0.001) to get the most of the best found parameters. It allows only to get to the best local converging optimum given the other parameters.


Potential extra idea could be adding an early.stop.round implementation.

It would be like this:
- compute the test data set
- use a watchlist using an eval_metric (can use logloss or any other evaluation metric if supplied as function...)
- set verbose = 0 to not flood the console (xgboost 0.47 -> xgb.train ignores verbose=0 due to a bug, but it's only for users who compile the latest xgboost version)
- set early.stop.round to an acceptable value (ex: 10), so it automatically stops once it converged to the test set

It would need a full rework of the functions though... xgb.cv is more appropriate also. It is possible to extract the performance metric list and the predictions from it (but not the model!!!). This way, it would automatically optimize for nrounds given the preset eta. Then, when the best parameters are found, run the model with the best parameters using xgb.train.


About:
# This is a toss-up, but the # trees probably adds complexity faster than number of splits

If it is about model complexity (getting to the local optimum of the provided dataset), in most cases: number of splits is faster to add complexity, but it is also prone to not hit the local convergence optimum when the split count is too large. Number of trees usually must get very large (like ~10000) when eta is very low (~0.001).
If it is about computation complexity (time to compute): have to assess it, number of trees is linearly dependent, unlike number of splits depending on colsample_bytree and overfitting control parameters.