consistently avoid dense matrix conversion for glmnet(x = ...)
currently, some of the functions for the glmnet model check if the
training data is a sparseMatrix, and some don't. the result is that
initial operations in train() might succeed, and then later in the
workflow, a step will fail (usually with "Cholmod error 'problem too
large'" for a sparseMatrix with very large dimensions) when some of the
training data is inadvertently converted to a dense Matrix. For
instance, this currently would happen when prob() in glmnet.R is called
(if trainControl(classProbs = TRUE)), or if tuneLength is used instead
of tuneGrid in train(), because tuneLength = ... triggers a call to
grid() in glmnet.R which does not do a check for a sparseMatrix.

this is essentially a more complete version of the fix for #1096